mk `fully_perform_op_and_get_region_constraint_data` a TypeOp method
